+++ src/main/java/com/hanuldure/project/dao/CommunityDAO.java
@@ -1,11 +1,22 @@
 package com.hanuldure.project.dao;
 
 import com.hanuldure.project.dto.CommunityTO;
-import java.util.List;
+import com.hanuldure.project.mapper.CommunityMapper;
+import org.apache.ibatis.session.SqlSession;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
 
-public interface CommunityDAO {
-    int insertCommunity(CommunityTO communityTO);
+@Repository("communityDAO")
+public class CommunityDAO {
 
-    CommunityTO selectCommunityBySeq(int board_seq);
+    @Autowired
+    private CommunityMapper mapper;
 
+    public int insertCommunity(CommunityTO communityTO) {
+        return mapper.insertCommunity(communityTO);
+    }
+
+    public CommunityTO selectCommunityBySeq(int board_seq) {
+        return mapper.selectCommunityBySeq(board_seq);
+    }
 }
